+++ CHANGELOG.md
@@ -8,10 +8,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 The following sections are considered for each release: **Added, Changed, Fixed, Security, Deprecated, Removed**
 
 ## [Unreleased]
+
+### Added
+- On task promotion, detect dangling reference set members and non-defining relationships left by Authoring Platform retire/delete actions and offer to tidy them (delete unreleased, inactivate released) before promoting (#1753)
+- When a task changes to status 'promoted' or 'completed' the associated ticket will now be automatically be changed to state 'Closed'
+- When a task is either deleted or has been released the ticket assocation to this task will now be automatically deleted.
+- The display of the tickets backlog page now includes the total number of tickets that is returned for a query.
+
+### Fixed
+- The Rows per page: 'All' option on the tasks page now correctly shows 1 - n of n instead of n + 1 - n  of n for the items displayed
+- Fix NMPC concept ids (e.g. Virtual medicinal product) leaking into the name generator input as raw SCTIDs by seeding the FSN/PT cache with `NmpcConstants` alongside `AmtConstants` and `SnomedConstants`.
 - Some concepts were defaulting to Primitive when they should not have been
 - After creating a task previously the application would sometimes hang for a long time, this no longer happens.
-
-
+- 
 ## [1.3.38] - 2026-04-24